post-review tidies and fixes
