+++ test/services/editions/publishing/PublishingTest.scala
@@ -7,7 +7,7 @@ import org.mockito.Mockito.{doNothing, never, times, verify, when}
 import org.scalatest.{FreeSpec, Matchers}
 import org.scalatestplus.mockito.MockitoSugar
 import services.editions.db.EditionsDB
-import services.editions.publishing.transform.FeastAppModel.FeastAppCuration
+import model.FeastAppModel.FeastAppCuration
 import services.editions.publishing.PublishedIssueFormatters._
 
 import java.time.LocalDate
@@ -16,15 +16,15 @@ import scala.util.{Failure, Try}
 class PublishingTest extends FreeSpec with Matchers with MockitoSugar {
   "publishing.publish" - {
     "should send Feast editions to Feast publication backend" - {
-      val editionsAppPublicationBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val editionsAppPreviewBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val feastAppPublicationTarget = mock[PublicationTargetBase[FeastAppCuration]]
+      val editionsAppPublicationBucket = mock[EditionsBucket]
+      val editionsAppPreviewBucket = mock[EditionsBucket]
+      val feastAppPublicationTarget = mock[FeastPublicationTarget]
       val db = mock[EditionsDB]
 
-      doNothing().when(feastAppPublicationTarget).putIssue(any,any)(any)
+      doNothing().when(feastAppPublicationTarget).putIssue(any,any)
       when(db.createIssueVersion(any,any,any)).thenReturn("new-version")
 
-      val toTest = new Publishing[PublishableIssue, FeastAppCuration](editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
+      val toTest = new Publishing(editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
 
       val iss = EditionsIssue(
         "fake-id",
@@ -44,21 +44,21 @@ class PublishingTest extends FreeSpec with Matchers with MockitoSugar {
 
       val expectedIssue = iss.toPublishableIssue("new-version", PublishAction.proof)
 
-      verify(editionsAppPublicationBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
-      verify(editionsAppPreviewBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
-      verify(feastAppPublicationTarget, times(1)).putIssue( org.mockito.ArgumentMatchers.eq(expectedIssue),org.mockito.ArgumentMatchers.eq(null))(any)
+      verify(editionsAppPublicationBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])
+      verify(editionsAppPreviewBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])
+      verify(feastAppPublicationTarget, times(1)).putIssue( org.mockito.ArgumentMatchers.eq(expectedIssue),org.mockito.ArgumentMatchers.eq(null))
     }
 
     "should send Editorial editions to Editions publication backend" - {
-      val editionsAppPublicationBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val editionsAppPreviewBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val feastAppPublicationTarget = mock[PublicationTargetBase[FeastAppCuration]]
+      val editionsAppPublicationBucket = mock[EditionsBucket]
+      val editionsAppPreviewBucket = mock[EditionsBucket]
+      val feastAppPublicationTarget = mock[FeastPublicationTarget]
       val db = mock[EditionsDB]
 
-      doNothing().when(editionsAppPublicationBucket).putIssue(any,any)(any)
+      doNothing().when(editionsAppPublicationBucket).putIssue(any,any)
       when(db.createIssueVersion(any,any,any)).thenReturn("new-version")
 
-      val toTest = new Publishing[PublishableIssue, FeastAppCuration](editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
+      val toTest = new Publishing(editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
 
       val iss = EditionsIssue(
         "fake-id",
@@ -78,23 +78,23 @@ class PublishingTest extends FreeSpec with Matchers with MockitoSugar {
 
       val expectedIssue = iss.toPublishableIssue("some-version-id", PublishAction.publish)
 
-      verify(editionsAppPublicationBucket, times(1)).putIssue( org.mockito.ArgumentMatchers.eq(expectedIssue),org.mockito.ArgumentMatchers.eq(null))(any)
-      verify(editionsAppPreviewBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
-      verify(feastAppPublicationTarget, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
+      verify(editionsAppPublicationBucket, times(1)).putIssue( org.mockito.ArgumentMatchers.eq(expectedIssue),org.mockito.ArgumentMatchers.eq(null))
+      verify(editionsAppPreviewBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])
+      verify(feastAppPublicationTarget, never()).putIssue(any[PublishableIssue],any[Option[String]])
     }
   }
 
   "publishing.updatePreview" - {
     "should send Editorial editions to Editions preview backend" - {
-      val editionsAppPublicationBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val editionsAppPreviewBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val feastAppPublicationTarget = mock[PublicationTargetBase[FeastAppCuration]]
+      val editionsAppPublicationBucket = mock[EditionsBucket]
+      val editionsAppPreviewBucket = mock[EditionsBucket]
+      val feastAppPublicationTarget = mock[FeastPublicationTarget]
       val db = mock[EditionsDB]
 
-      doNothing().when(feastAppPublicationTarget).putIssue(any,any)(any)
+      doNothing().when(feastAppPublicationTarget).putIssue(any,any)
       when(db.createIssueVersion(any,any,any)).thenReturn("new-version")
 
-      val toTest = new Publishing[PublishableIssue, FeastAppCuration](editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
+      val toTest = new Publishing(editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
 
       val iss = EditionsIssue(
         "fake-id",
@@ -114,21 +114,21 @@ class PublishingTest extends FreeSpec with Matchers with MockitoSugar {
 
       val expectedIssue = iss.toPreviewIssue
 
-      verify(editionsAppPublicationBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
-      verify(editionsAppPreviewBucket, times(1)).putIssue( org.mockito.ArgumentMatchers.eq(expectedIssue),org.mockito.ArgumentMatchers.eq(null))(any)
-      verify(feastAppPublicationTarget, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
+      verify(editionsAppPublicationBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])
+      verify(editionsAppPreviewBucket, times(1)).putIssue( org.mockito.ArgumentMatchers.eq(expectedIssue),org.mockito.ArgumentMatchers.eq(null))
+      verify(feastAppPublicationTarget, never()).putIssue(any[PublishableIssue],any[Option[String]])
     }
 
     "should do nothing if you pass a Feast edition" - {
-      val editionsAppPublicationBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val editionsAppPreviewBucket = mock[PublicationTargetBase[PublishableIssue]]
-      val feastAppPublicationTarget = mock[PublicationTargetBase[FeastAppCuration]]
+      val editionsAppPublicationBucket = mock[EditionsBucket]
+      val editionsAppPreviewBucket = mock[EditionsBucket]
+      val feastAppPublicationTarget = mock[FeastPublicationTarget]
       val db = mock[EditionsDB]
 
-      doNothing().when(feastAppPublicationTarget).putIssue(any,any)(any)
+      doNothing().when(feastAppPublicationTarget).putIssue(any,any)
       when(db.createIssueVersion(any,any,any)).thenReturn("new-version")
 
-      val toTest = new Publishing[PublishableIssue, FeastAppCuration](editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
+      val toTest = new Publishing(editionsAppPublicationBucket, editionsAppPreviewBucket, feastAppPublicationTarget, db)
 
       val iss = EditionsIssue(
         "fake-id",
@@ -148,9 +148,9 @@ class PublishingTest extends FreeSpec with Matchers with MockitoSugar {
 
       assert(result.isSuccess)
 
-      verify(editionsAppPublicationBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
-      verify(editionsAppPreviewBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
-      verify(feastAppPublicationTarget, never()).putIssue(any[PublishableIssue],any[Option[String]])(any)
+      verify(editionsAppPublicationBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])
+      verify(editionsAppPreviewBucket, never()).putIssue(any[PublishableIssue],any[Option[String]])
+      verify(feastAppPublicationTarget, never()).putIssue(any[PublishableIssue],any[Option[String]])
     }
   }
 }
